+++ lib/consts.tsx
@@ -16,3 +16,5 @@ export const DEFAULT_FRAME = {
 } as FrameMetadataType;
 
 export const SOUND_FACTORY = '0x0000000000aec84F5BFc2af15EAfb943bf4e3522';
+
+export const AIRSTACK_API_URL = 'https://api.airstack.xyz/graphql';
